+++ python/__init__.py
@@ -1,0 +1,6 @@
+''' This module is the main module for the python package. '''
+
+from . import database
+from . import scripts
+from . import solver
+from . import utils
