+++ python/database/__init__.py
@@ -1,6 +1,7 @@
 ''' This file is used to import all the models from the models.py file. '''
 
-from .models import *
-from .db import *
+from . import imports
+from . import models
+from . import query
 
 from . import sqlalchemy_main
