+++ python/database/models/__init__.py
@@ -1,10 +1,7 @@
 ''' SQLAlchemy Model Classes'''
 
-from .model_config import ModelRun, ModelConfig
-from .edes import EDES
-from .precinct_distance import PrecintDistance
-# from .model_run import ModelRun
-from .residence_distance import ResidenceDistance
-from .result import Result
-
-
+from .model_config import *
+from .edes import *
+from .precinct_distance import *
+from .residence_distance import *
+from .result import *
