+++ python/database/models/edes.py
@@ -5,7 +5,7 @@
 
 from sqlalchemy import Column, String, Integer, Float, text
 
-from python.database.sqlalchemy_main import ModelBase
+from ..sqlalchemy_main import ModelBase
 from python.utils import generate_uuid
 
 class EDES(ModelBase):
