+++ python/database/models/model_config.py
@@ -8,7 +8,7 @@ import json
 from sqlalchemy import Column, String, Integer, Float, Boolean, DateTime, ForeignKey, text, ARRAY
 from sqlalchemy.orm import mapped_column, Mapped, relationship
 
-from python.database.sqlalchemy_main import ModelBase
+from ..sqlalchemy_main import ModelBase
 from python.utils import current_time_utc, generate_uuid
 
 class ModelConfig(ModelBase):
