--- python/scripts/__init__.py
@@ -1,5 +1,0 @@
-''' This file is used to import all the scripts in the scripts folder. '''
-from . import address_validator
-from . import auto_generate_config
-from . import db_import_cli
-from . import model_run_db_cli
