+++ python/solver/__init__.py
@@ -1,6 +1,5 @@
 ''' This module contains the files required to run the optimization solver. '''
-from ..utils.constants import PROJECT_ROOT_DIR, DATASETS_DIR, RESULTS_BASE_DIR
-from model_config import PollingModelConfig, MODEL_CONFIG_ARRAY_NAMES
+
 from . import (
   model_data,
   model_factory,
