+++ python/utils/__init__.py
@@ -1,6 +1,7 @@
 ''' This module contains utility functions for the project. '''
 from .utils import *
 
+from . import constants
 from . import csv_constants
 from . import pull_census_data
