+++ requirements.txt
@@ -1,5 +1,6 @@
 sqlalchemy-bigquery==1.12.0
 alembic==1.13.3
 google-cloud-bigquery-storage==2.27.0
+googlemaps==4.10.0
 pyarrow==18.0.0
 tqdm==4.67.0
