+++ tests/test_auto_generate_config_1.py
