+++ tests/testing_config_penalty.yaml
